I created  a better popup window for when the user wants to exit the application without saving first. It now shows up *before* the app would close, and all 3 buttons are working as intended.
If the user closes the popup window with the default windows X icon, 
then it doesn't save the output and the entire app closes.
I will later modify it to be equivalent of the "cancel" button

Right now the popup will appear regardless if the user saved before
wanting to close the app or not. I've created a method that compares
the contents of the two output files and if they are the same, the
popup should not appear, however if I include it in the current 
confirmCloseRequestHandler it crashes the app for an unknown reason.